+++ src/components/SignupForm.vue
@@ -24,7 +24,7 @@
 </template>
 
 <script>
-import { registerUser } from '@/api/index';
+import { registerUser } from '@/api/auth';
 import { validateEmail } from '@/utils/validation.js';
 
 export default {
